Move tests to gravy helpers tests
